+++ packages/base/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [1.1]
+### Added
+- Bump `env` from `es2020` to `es2022`
+
 ## [1.0]
 ### Added
 - Initial version