+++ packages/typescript/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [2.0]
+### Breaking
+- Bump `@typescript-eslint` from `v5` to `v6` and enable strict rules
+
 ## [1.0]
 ### Added
 - Initial version