+++ packages/vue/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [2.0]
+### Breaking
+- Enforce `kebab-case` casing for custom events with `vue/custom-event-name-casing` rule
+
 ## [1.0]
 ### Added
 - Initial version